FAB-1783 - use installed CC from file system
https://jira.hyperledger.org/browse/FAB-1783. In particular,
"newdeploy.pptx" there.

Move off per channel "on-ledger" deploy model to "on-peer" install
model. This affects version, launch and per channel ledger instatiation.

Now that chaincode is not on ledger but is on-peer, there is no need
to create a container per instance of the chaincode deployed on
different channels. These are the tightly related changes
   . user has to specify version (see, version option in CLI commands)
   . load off file system and not on ledger (see ccprovider.go)
   . chaincode launch separated from chaincode instantiation (see chaincode_support.go)
   . chaincode "init" is no longer tied to deploy (install) (see handler.go)

Note 1 -a minor side effect is that the chaincode container is not killed after instantiate
like it was after the old deploy. It need not be as the chaincode is not bound to a
channel.

Note 2 - the "--peer-chaincodedev=true" has to be treated differently for the
new model. A JIRA https://jira.hyperledger.org/browse/FAB-2128 has been opened
for this

Change-Id: I34b417d119c16bf296130f8740cddd5af9a8b582
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>