Re-enable CAR test
We upgrade to the v0.10.1 release of chaintool which properly
supports the latest fabric.  This allows us to re-enable the
car unit-test to ensure things do not break in the future.

We also add a feature which allows us to overide the version
of chaintool at the command-line, which was helpful in
formulating the development of the fixes for chaintool as
well as the ultimate release of v0.10.1.  To use, simply
set $CHAINTOOL_URL on the command-line for the build.

Helpful tip: it supports file:/// schema.

Change-Id: I952e147d9bfe7836517cb5b08f42a12493be05b6
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>