Decouple gossip.LeadershipMsg and networkMember
The proto message that is used for leader election
contains Member, which contains endpoint, metadata and pkiID.
There is no use of endpoint in the leader election module,
and also- the use of metadata is abuse since its a []byte,
but all we do is to put there a bool.
I replaced the use of Member by pkiID and a bool value
that was previously in the metadata.

Change-Id: Id31395c5978f6b8e1791bad35d28a8a66e6cf664
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>