[FAB-1411] Pull the zookeeper tarball by curl
Pull the zookeeper tarball by using curl instead of wget, so it can
understand the "socks*://" style proxy protocol prefixes, yet support
wget's "http*://" scheme.

Not exactly the FAB-1411 but closely relate to.

Extract the archive using a pipeline, which avoid leaving the allocated
space of the archive as dark matter in the image's layers (and reduce
code length BTW).

Change-Id: I2cf8f04f7ba98c99ceeb2befe8d92c78b8b9e332
Signed-off-by: Ray Chen <ray@hyperchain.cn>