[FAB-1984] Remove GetQueryResult from QSCC
In order to control access, all queries will need to go through
application chaincode. This change is in response to early
v1 feedback.

See marbles02 for an example of calling GetQueryResult from application chaincode.

In the future if there is capability to place ACL on peer APIs,
then it would be possible to expose GetQueryResult as a peer
API (or via QSCC).

Change-Id: I1613563caf823e45fdfebeaa133f4ec523716cec
Signed-off-by: denyeart <enyeart@us.ibm.com>