[FAB-2198] Introduce envelopes to gossip message
The current gossip implementation uses proto.Marshal
and then signs over this output both in the signer
and the verifier, which might be dangerous because protobuf
marshalling isn't guaranteed to be deterministic.

This is the first commit in the series that introduces
an envelope message to the GossipMessage, and does
some refactoring and preperation for next commit(s)
that'll change the stream definition, the signing,
and entity-storage related structures.

The refactoring basically entails moving the ReceivedMessage
from comm to proto, as part of the extensions to the gossip proto
types.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>
Change-Id: Ifcedaa87bcf45376a4569ed80b7e7d1b76c87883