FAB-2203 handle chaincode launch serially
https://jira.hyperledger.org/browse/FAB-2203

Chaincode-peer communication was made asynch to handle concurrent
invokes. The state machine requires that launch (one time operation)
requires serial communication to get into "ready" state. While it is not
incorrect to have async communication for launch also, it makes it hard
to write test cases against (we have to wait for CC to get to ready state).

A better approach would be for the system to enforce that at the end of init
the system woukd be in ready state for inits/invokes. This is achieved by
sending ready state serially to the CC.

Change-Id: Ib5d3ec1b6188d5aa933dbcb43adaf9995f75de2d
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>