[FAB-2312] configtx value handlers to own package
https://jira.hyperledger.org/browse/FAB-2312

There used to be a single type for configuration, 'ConfigItem', but that
has bifurcated into 'ConfigValue', 'ConfigGroup', and 'ConfigPolicy'.
In order to handle the 'ConfigValue' and the 'ConfigPolicy' in similar
ways, the 'ConfigValue' handling needs to be pushed out into its own
package just like policy handling is.  This CR accomplishes that.

This is in preparation for treating the PolicyHandler as more of a first
class object and to eliminate the need for hacky importing.

Change-Id: I690a33e09843eb4c29611319f2e3942baae2dfdb
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>