[FAB-2322] Allow mod_policy to be relative
https://jira.hyperledger.org/browse/FAB-2322

With the notion of hierarchical policies, the configuration mod_policy
needs to be able to refer to policies in the current group (relatively)
or policies relative to the channel root configuration (absolute).

This CR modifies the policy lookup to ask the relative policy manager
unless the policy begins with a "/" in which case the policy is
retrieved from the root policy manager.

Change-Id: Id14e327a5c37c4c1848521abd691aa857a12039d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>