TOC Proposal
This is a proposed re-architecture of the community
documentation. IT IS NOT PERFECT.
Do note that the majority of these
new files are simply placeholders for in progress
work. Questions to answer - does the flow/layout
make sense? - are there missing topics? - are the
topic headings intuitive?
Disregard the formatting (e.g. inability to collapse
sections, redundant headings, etc. That will be
addressed in another patch set).
[ci skip]

Change-Id: Ic1ac43adc7c40c0be12355fedf35a42417716e26
Signed-off-by: Nick Gaski <ngaski@us.ibm.com>