Remove dup Transaction timestamp and version fields
The Transaction timestamp and version fields are duplicates
of fields that are already on the transaction's ChainHeader, and
nobody is using them.  The header is the appropriate place for these
fields, not in the Transaction itself.  Note, there is a 1-to-1
relationship between the ChainHeader and the Transaction, no
reason to include the fields twice.

Change-Id: I039d4ca07d51d27d0e9c4b1a00f9b39db2ba08c8
Signed-off-by: denyeart <enyeart@us.ibm.com>