[FAB-2339] Add simple tool write out genesis block
https://jira.hyperledger.org/browse/FAB-2339

In order to simulate real deployment scenarios, it's necessary to have
the genesis block generated ahead of time and loaded via a file.  This
has been done only via bdd thusfar.  This CR adds a simple tool which
dumps the genesis block created by the configtx provisional genesis
package to the filesystem.

There is a doc file under docs/configtxgen.md which describes the usage.

Change-Id: I42171afdce4bb47d63806805a88aa3b93759023d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>