FAB-2357 join failed looking up ledger
https://jira.hyperledger.org/browse/FAB-2357

Not all proposals are on channels (and have a ledger). Some proposals
such as to the CSCC for join is "channelless".
In this instance, the failure is due to attempting to access the ledger
for doing TX uniqueness checks. Given there's no ledger, we have to
bypass ledger based uniqueness checks.

Change-Id: I3dbdf49307989575494efc15f3586bd4a85c41cb
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>