[FAB-2408] Fix policies absolute paths
https://jira.hyperledger.org/browse/FAB-2408

While the policy manager was still in development, the absolute path
resolution for policies was handled in the configuration manager.  This
worked fine for config updates, but for other components of the system
such as those depending on the 'ChannelApplicationReaders', the absolute
paths did not work.

This CR pushes the absolute policy path resolution back into the policy
manager where it belongs.

Change-Id: Idcfa2ca030115b3ff9e3034f8c54767af1c2ea49
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>