[FAB-2424] Enforce MSP channel validation in gossip
Before this commit, peers of an organization could publish assertions
for having joined a channel, and it was enough for them to obtain blocks
from peers in the same organization.

This commit integrates gossip with the MSP method (VerifyByChannel) that consults
the MSP policies whether a certain peer is indeed eligible for receiving blocks
for the channel.

I changed the gossip tests and the channel tests to mock the method,
and also added a test in the channel test that checks that even when we query
the membership of a channel, it also calls VerifyByChannel.

I also had to make the MembershipStore thread safe for the channel

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>
Change-Id: I0fccf98080d7e72d05e7e762244fe366e9f6e32a