Fix break in master
When a new channel is created with `peer channel create`, one extra org is
generated for gossip. This extra org brings the total number of orgs to 2 and
requires a majority of 2 when admin policies are evaluated. However, this
extra org does not come with its admin policy and so the reject policy is
used instead. This breaks `peer chaincode instantiate` because the policy
evaluation fails. The fix is to add a permit-all policy for the extra org.

Change-Id: Ia35da0d83bea7c9ef8fe785164d1daa158791969
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>