Add query to get the installed chaincodes on a peer
This change adds a query which returns the details of all
chaincodes installed on a peer. The return value is a
ChaincodeQueryResponse proto which contains an array of
ChaincodeInfo protos, which each contain the chaincode name,
version, and path (the input arguments, ESCC name, and VSCC
name will all be blank since those fields are  info only
related to instantiating a chaincode).

https://jira.hyperledger.org/browse/FAB-2235

Change-Id: I5154678729211f62c1f686c99dafa598613e49ea
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>