Add query to get all the channels for a given peer
This change adds a query which returns information about
all the channels for a given peer. The return value is a
ChannelQueryResponse proto which contains an array of
ChannelInfo protos, which each contain the channel id.
In the future, more information about the channels will
likely be added to the ChannelInfo proto.

https://jira.hyperledger.org/browse/FAB-2237

Change-Id: I46b96b37cfa607b7996fa65790e3288e0d0983b2
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>