Gossip- Don't log sending to empty slice of peers
The code that uses the gossip comm layer's Send() method
computes dynamically a set of peers to send a message to,
usually out of a batch of all kinds of messages.

As a result, the peer set passed to the send method is sometimes empty.
This makes the gossip comm layer log something like "sending Msg M to 0 peers".

I added a check that makes the Send() function skip sending if the peer slice
passed is empty.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>
Change-Id: I689f3163f496ac91629536e63a16d0e7effd99e3