[FAB-2473] Fix gossip proto style 3/3
https://jira.hyperledger.org/browse/FAB-2473

This is the third and final in a series of changesets to move the gossip
protos to conform with the official proto style guidelines.

This CR is the most invasive, because it changes pkiID to be pki_id,
causing the generated code to become PkiId instead of PkiId.  This is
very similar to the change from ChaincodeID to ChaincodeId, but much
smaller in scope.

Change-Id: I7ebd2db48ef464ac7f1d024a41fbd11514ca5385
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>