Gossip- warn and abort on invalid join channel event
When the MSP-ID of the peer isn't among any of the names of the organizations
that are members of a channel, the gossip layer of the peer doesn't see
any members of its own org as members of the channel.

It doesn't make any sense to have a peer join a channel it isn't
a member of, and this scenario confuses the gossip layer of the peer.

Therefore, I am adding a check that prevents the gossip layer
from joining a channel in such a case, and logs an error.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>
Change-Id: Ida456f72b1ab6813fbdcfdb0393b8f0eb7b516e7