[FAB-2511] Make configtx sequence explicit
https://jira.hyperledger.org/browse/FAB-2511

The previous non-MVCC configuration system required that the modified
version numbers of the configuration elements be equal to the global
configuration sequence lock.

This makes the global sequence explicit, while allowing the individual
elements to move (or not move) their version numbers locally.  This
finally enables the partial update functionality advertised in FAB-1880.

Change-Id: Ic2f554b864e91435a8812d56eb8247cc1fa33720
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>