[FAB-2552] Allow concurrent config proposals
https://jira.hyperledger.org/browse/FAB-2552

The original config proposal framework was designed under the idea that
only one committable proposal was ever given at a time.  However,
because of the need to support partial updates, there are additional
situations where a commit-like path needs to be performed for the config
concurrently with other commit-like paths.

This CR adds an internal tx identifier to configuration proposals to
allow for the safe concurrent processing of configuration.

Change-Id: I9f8cb5eb8a6c426fb930591792781e1159f8e601
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>