[FAB-2554] configtx.Manager track deserialized val
https://jira.hyperledger.org/browse/FAB-2554

The configtx Manager currently calls into the config deserialization but
does not retain the unmarshaled config values (as they are retained and
accessible through the config).

In order for the configtx manager to also support debugging and
inspection of configuration, it is necessary that these deserialized
versions be tracked by the manager. This CR adds this functionality.

Change-Id: I37165e8d5877dd06dc29fad95831069cf7f41d30
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>