FAB-2531 Range queries fail iterating beyond 100 items
This is a temporary stabilization change until the following issue is
resolved and will prevent errors in queries for the beta:
FAB-2462 - Re-introduce paging for range queries and rich queries

Peer sends 100 results to shim.
When shim asks for 101st item, the request for the next 100 goes
back to peer and it fails due to a closed iterator.

This change also sets the HasMore flag on query result sets to
false.  This will prevent the requery from being called by the
shim and throwing an exception.

Purpose of this change is to stabilize the behavior for the
beta by setting a query size limit in core.yaml at 10000.

Then use the query size limit in the CouchDB query and in
the chaincode handler.

Added changes to chaincodetest.yaml to match core.yaml.

Added unit test cases for the new querylimit option.

Change-Id: I772d1f87beec2296db2eed68a0528181ac1ddeca
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>