[FAB-2574] Config parsing outside configtx.Manager
https://jira.hyperledger.org/browse/FAB-2574

The configtx.Manager is currently the only place which parses a
configuration transaction (appropriately unmarshaling all of the
config values).

In order to facilitate the inspection of configuration
transactions outside of the context of a full running system, this
CR splits the configuration parsing from the configtx.Manager (although
the configtx.Manager is the only consumer of this code).

Change-Id: I91d7fb880892f4e88cf50bc3ba930f67204dcc52
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>