FAB-2524 - instantiate/upgrade should not specify code
https://jira.hyperledger.org/browse/FAB-2524

Instantiate and Upgrade continue to use the ChaincodeDeploymentSpec.
They should instead use the CDS from the installed package.

We should note that a better way to do that is to not overload
ChaincodeDeploymentSpec but use it only for the install command. That is
a much bigger change involving protocol and best done in a future CR.

We could check instantiate and upgrade calls just have minimal
information in the CDS and return error. This is not done in this
CR for two reasons
   . this is a tactical fix pending the right protocol change mentioned
   . don't want to break SDKs which may be sending those spurious
     information. We will just ignore such information. They can make
     the change at their convenience

Change-Id: I9ee8e6e347e3f935626d8b113dbe512a85859f19
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>