[FAB-2589] Clean up top-level orderer pkg
https://jira.hyperledger.org/browse/FAB-2589

This changeset:

1. Switches all panic calls to their logging-enabled equivalents.

2. Converts all logger calls that take a format specifier but print a
single argument at the end of the string to their no-format equivalents.
There is no reason to use `logger.Errorf("Failed to listen: %s", err)`.
`logger.Error("Failed to listen:", err)` returns the _exact_ same output
(including an automatically inserted space before the argument).

3. Remove unnecessary log-level setting call in the beginning of the
`main()` function. If the call to `InitFromSpec` is fed with junk, it
will default to the INFO level anyway.

Change-Id: Iaaa6047305bd8c1d10e580fca42a88181d16791e
Signed-off-by: Kostas Christidis <kostas@christidis.io>