Fix peer panic at initiate proposal with no MSPs
This change ensures the peer doesn't panic when a proposal
is initiated and no MSPs are defined. This is due to the
code checking the outer MSPConfigHandler structure for nil
instead of the MSPManager interface inside it.

It also updates protos/utils/proputils_test.go to a blackbox
test to avoid an import cycle.

https://jira.hyperledger.org/browse/FAB-2536

Change-Id: I61b0366a127b4d90e2b187d2dc0e91a5541a243c
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>