[FAB-2576] Mock committer ledger height
The reason mockLedgerInfo should be initialized with 1 instead of 0, is
because the state transfer expected to get the ledger with at least one
block committed. Which is the case in reality, since to create a ledger
you need to receive genesis block and genesis block committed right
away.

Without this assumption since blocks sequence uses uint64 there is an
overflow to -1 which cause anti entropy procedure to request big bulk of
blocks and extensive memory footprint.

Change-Id: Id836b20d8be474a5fef492d3082f5878e87aa738
Signed-off-by: Gennady Laventman <gennady@il.ibm.com>
Signed-off-by: Artem Barger <bartem@il.ibm.com>