[FAB-2632] Default endorsement policy
If no custom endorsement policy was specified, the peer used to set a default
endorsement policy requiring a signature from a member of MSP "DEFAULT".
However this MSP may not always exist. The solution is to default to the
following policy: a signature from a member of any of the application MSPs
specified for the channel.

Change-Id: Ie3d5876cc8593b5b91babb4ba21f3146338cd3bd
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>