[FAb-2680] Set default ModPolicy in configtxgen
https://jira.hyperledger.org/browse/FAB-2680

The ModPolicy field of the configuration elements is currently set to
the empty string.  This causes all configuration updates to be denied by
default (As there is no policy with name empty string).

This CR enhances configtxgen to set the default modification policy of
"Admins" for all items.  This is a sane default, and there is currently
no mechanism to change it.

This is a very low risk change from a regression standpoint, because it
only affect configuration updates (which are currently exercised nowhere
in the system).

Change-Id: Ica16c0a56f8cd42e5f02c770f0f10c59328b63d7
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>