[FAB-2696] Default chain broken in peer
https://jira.hyperledger.org/browse/FAB-2696

According to @aso, the default chain is currently broken because it does
not appropriately encode the channel config, and additional config
sanity checks are now rejecting the config as invalid.

This CR switches the default chain to using the configtxgen supporting
structures to generate the genesis block for the default chain, so that
all of the config checks will pass.

It also causes the configtxgen tool to search a few different paths for
relative MSP directories (., PEER_CFG_PATH, ORDERER_CFG_PATH, and
finally the GOPATH).

Change-Id: Ib6b836212e335b57f0c2b3c23dc4f05e9a47c56e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>