[FAB-2773] Restrict the total count of channels
https://jira.hyperledger.org/browse/FAB-2773

There is a risk that a malicious user with channel creation rights could
simply create an unlimited number of channels until exhausting the
resources of the ordering network, causing it to fail.

This CR introduces a new genesis block parameter ChannelRestrictions
which for now, contains only max_count.  This parameter is checked
before accepting a new channel creation request.

Note that this must be a genesis parameter so that it is synchronized
across the network and all orderers agree to accept or reject channel
creation requests deterministically.

Change-Id: I582963925e05d50e6bad6d1aa7d2a490c0e45cc1
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>