[FAB-2821] Print readset/writeset/deltaset on updt
The configtxgen tool currently only prints the organizations in the
channel creation configtx.  This is not enough information for an admin
to decide whether to sign or not.

This CR makes the configtx printing much verbose, including the read
set, write set, and computes the delta set for easy identification of
non-overlap between the read and write sets.

Change-Id: I3f9bfc2839ec595b1f2fc9a0cb149797ed4f862b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Signed-off-by: Kostas Christidis <kostas@christidis.io>