[FAB-2901] Fixing election test
During latest changes leader election callback invocation
was changed from sync to async (go routine). It caused to tests
that check if callback was invoked run before it was actually
invoked. Test changed to handle this - if result is not as expected,
we wait for a while and check again.

Change-Id: Ib26dc5c7d48110c36d39a216e83bcc4a99981389
Signed-off-by: Gennady Laventman <gennady@il.ibm.com>