[FAB-3000] Remove reliance on vagrant baseimage
Maintaining the vagrant-baseimage is difficult since we lack
suitable automation facilities.  It has therefore become an
impediment to producing new docker baseimages.  In addition,
it is also less important these days since most of the heavy
lifting is performed by the docker infrastructure.

Consider that when the baseimage for vagrant was created, we
still needed to compile a bunch of basic tools (golang, protobufs
rocksdb, etc).  This is no longer true: Some of the tools
are now available in binary form, others are exectuted within
docker, and others are no longer used (e.g. rocksdb).

Therefore, we can greatly simplify our processes by simply
eliminating our reliance on a precompiled baseimage for
vagrant.  Because of the dockerization effort, all we really
need is a basic platform with make, git, golang-1.7, and docker.

We also include a few other items for developer convenience,
such as behave, nodejs, and java.

Fixes FAB-3000

Change-Id: Ic0749a1c0361a0b4a83d7ca6879e9d2bb50a5456
Signed-off-by: Gregory Haskins <gregory.haskins@gmail.com>