[FAB-3012] Copy callstack from lower level error
This change adds the ability to copy the callstack and error message
from a lower level error into a later error. For example, an error may
occur in a common utility function that a different package may want to
wrap with its own component/reason code, error message, and callstack
while retaining the details from the wrapped error.

Change-Id: Id143640bbaaba7bf125b343dbc614d046acbfd1a
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>