Cleanup: removing accesscontrol package
This change-set does the following:

1. It removes the fabric/accesscontrol package.
The reason for removing that package is because
it is not compatible anymore with v1.
A side effect of this is that the chaincode examples
using that package have been removed as well.
At the current stage, a chaincode developer can perform
access control by reasoning about the creator of a proposal.

2. It removes the fabric/core/crypto package.
The accesscontrol package was the only one using
that package. Recall that all the crypto operation
are now performed using the BCCSP.

Change-Id: If222d353974cba1ec1bd7f5d5dc5451f9a455c18
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>