[FAB-2927] install chaincode package
The CLI install command is enhanced to install a package
previously generated with the "package" or "signpackage" commands.

The cc-packaging-and-signing.rst describes the command.

With this change chaincode can deployed in one of these methods
  . using SDK or older "install" commands
  . create a (signed) package first as described in
    cc-packaging-and-signing.rst

fabric will first test for the older "ChaincodeDeploymentSpec" based
install before attempting to use the package.

Security fixes and linking the install to the instantiation record will
be submitted in future CRs.

Change-Id: I90c736d5f933dcd619b4de5944e7b69b50b45893
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>