[FAB-3094] CouchDB Errors if system databases missing
CouchDB log fills with errors if its system databases are not created.
The system databases should be created upon peer startup,
if they do not yet exist.

This change makes a simple check on creating the state database to
check if the system database exist as well.  If they are missing, then
the system databases are created.

Change-Id: I6dacb8ceefcda8fd0f14a0ef55da921b9160343d
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>