[FAB-3009] Clean up core.yaml logging section
This CR handles a number of core.yaml cleanup operations for the
logging section. It removes unused parameters, adds new parameters for
the 'gossip' and 'ledger' modules (that will be utilized in a future
CR), and updates the comments (which were outdated). It also ensures
that 'logging.peer' is used to set the default log level when the
CORE_LOGGING_LEVEL environment variable is not set.

Change-Id: Ib94d4b991570536b4fb195b472f381c565fd13ba
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>