[FAB-3112] Do not include configtx helper.go at runtime
The current code assumes, by virtue of the init() routine in helper.go,
that an MSP configuration that is compatible with "DEFAULT"/"sampleconfig"
will exist in all contexts (test, development, and runtime).  While this
is true for test and dev, runtime is not likely to have the sampleconfig
available (nor should it).  Therefore, we fix the assumption by removing
the implicit init() code and only use the logic on demand.

It should be noted that any attempts to run "peer channel create" without
any file as input will panic under a "runtime" environment (e.g. installed
on a system, not built from source).  This is true both before and after
this patch.  This should probably be fixed as a separate issue.

Change-Id: I33bcd7d5be716a695c65fc09ae96807a2769f1d9
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>