[FAB-3136] Include install script with release package
Since we distribute our runtime(s) as Docker images
and there are several of them to download, it's
convenient to provide a script which will download
the correct version as well as iamges for the correct
architecture.

The script basically has a list of the various fabric
images and the make target uses substituion to set
the right variables for each platform.

NOTE:  I hard-coded VERSION to 1.0.0-alpha for now
because we do need to publish these packages and
update the docs to ensure things will still work with
the alpha

Change-Id: I76d6db3ab638e13e65c4b4c0ce23d4af4e6ef068
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>