[FAB-2061] Gossip inter-org confidentiality - P1
This change set makes every 2 orgs (that share a channel),
gossip only alive messages that that originated in one of the two
organizations.

The motivation is to prevent a third organization from knowing
that the two orgs share a common channel (unless of course, they all share it).

So, if we have the following situation:
Channel C0: { orgA, orgB }
Channel C1: { orgB, orgC }
[ A ]--C0--[ B ]--C1--[ C ]
Then orgA has no idea of orgC's existence and vice-versa.

But, also if we have the following situation:
Channel C0: { orgA, orgB }
Channel C1: { orgA, orgC }
Channel C2: { orgB, orgC }
[ A ]-C0-[ B ]
  |      /
  |     /
  C1   C2
  |   /
  |  /
[ C ]

Then, alive message from orgA, are never sent
to orgB via orgC (and the same for orgB, orgC)
Or, in other words: an Alive message m about a peer in org X
is sent to org Y, only if X is either in X or in Y.

In the next change set I'll do the same for identity messages
(messages that contain certificates)

Change-Id: I816e6b23050d320f8b99a64b5c84d619de54aaa8
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>