[FAB-2061] Gossip inter-org confidentiality - P2
In the previous commit, I adjusted the membership routing so that
membership would only be disseminated between pairs of organizations.

The only thing left is to deal with identity messages, which are
disseminated using the pull mechanism.

This commit adds a filtering capability to the pull engine,
and to its enclosing object- the pull mediator.

The filter acts on a context of a message (that contains info about the
peer's credentials, hence it's organization) sent from the remote peer
(helloMsg and requestMsg) and return a function that declares for each
item (digest or item in the response) whether it should be sent to the remote peer.

In the next commit, I shall write filtering logic to make it so
that peers would gossip identities with destination orgs only
if they are their own identities.

Change-Id: I7996bafe5ce2962279c01e59c57868057cfd1d7e
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>