[FAB-2963] Gossip inter-org confidentiality - P4
In gossip, AliveMessages of peers that have no external endpoints
are only gossiped among peers of the same organization.
However, no such restriction exists for StateInfo messages and for
Identities. Therefore- foreign organizations know of peers
that they can't communicate with, and that the organization had no
intention of exposing.

This commit addresses this by filtering out StateInfo messages
of peers without external endpoints when gossiping to peers of foreign
organizations, and the same for identity messages.

Change-Id: Idf7953b9fa5efd1415ac294ff63e7ea321762a3f
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>