[FAB-2499] different OrgID and OrgName in configtx.yaml
Previously, in order for gossip to be able to join a channel,
the OrgID (that sets the MSP-ID) and the OrgName in the configtx.yaml
were needed to be the same.

This commit removes this requirement by using the fact that ApplicationOrg
embeds Org and the latter has an MSPID() method.

I tested this by using a configtx.yaml with different OrgName and OrgID.

Change-Id: I6435047fa97b16532491a22490dcde0f151945b0
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>